Extract playlist core infrastructure
